cr fixes
