+++ src/apps/ScopeChangeRequest/Components/Form/Origin.tsx
@@ -79,6 +79,5 @@ const Wrapper = styled.div`
     display: flex;
     align-items: flex-end;
     justify-content: space-between;
-    width: -webkit-fill-available;
     padding: 0.14em 0em;
 `;
